# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Boop API Example",
    "version": "1.0"
  },
  "paths": {
...
    "/api/news": {
      "get": {
        "summary": "Get all news articles.",
        "tags": [
          "News"
        ],
        "responses": {
          "200": {
            "description": "A list of news articles.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/News"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      },
      "post": {
        "summary": "Create a new news article.",
        "tags": [
          "News"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created news article.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
    "/api/photo": {
      "get": {
        "summary": "Get All Photo",
        "tags": [
          "Photo"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PhotoResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cannot fetch"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add new Photo",
        "tags": [
          "Photo"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Photo added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "News": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "location": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageId": {
            "type": "string"
          },
          "sourceLink": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "titleUA": {
            "type": "string"
          },
          "textUA": {
            "type": "string"
          },
          "titleEN": {
            "type": "string"
          },
          "textEN": {
            "type": "string"
          },
          "titleIT": {
            "type": "string"
          },
          "textIT": {
            "type": "string"
          }
        }
      }
    }
  },
...
}